- merging the two versions of PhysicalLocation.  Now there are two concrete implementations, one that stores x/y as shorts, and the other as ints.  The former has the default behavior of overlfowing with read names coming from certain versions of Illumina Software, while the latter currently does not.  Leaving for future work, clean up of the various physical location like classes.
