+++ src/java/picard/sam/markduplicates/util/AbstractMarkDuplicatesCommandLineProgram.java
@@ -40,6 +40,7 @@ import picard.PicardException;
 import picard.cmdline.Option;
 import picard.cmdline.StandardOptionDefinitions;
 import picard.sam.DuplicationMetrics;
+import picard.sam.util.PhysicalLocation;
 
 import java.io.File;
 import java.util.ArrayList;
@@ -274,7 +275,7 @@ public abstract class AbstractMarkDuplicatesCommandLineProgram extends AbstractO
      * optical duplicate detection, we do not consider them duplicates if one read as FR ann the other RF when we order orientation by the
      * first mate sequenced (read #1 of the pair).
      */
-    private static void trackOpticalDuplicates(final List<? extends OpticalDuplicateFinder.PhysicalLocation> list,
+    private static void trackOpticalDuplicates(final List<? extends PhysicalLocation> list,
                                                final OpticalDuplicateFinder opticalDuplicateFinder,
                                                final Histogram<Short> opticalDuplicatesByLibraryId) {
         final boolean[] opticalDuplicateFlags = opticalDuplicateFinder.findOpticalDuplicates(list);
